# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API for Big Bad Con",
    "version": "1.0",
    "title": "Big Bad Con data service REST API",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Jeremy Tidwell",
      "url": "https://www.bigbadcon.com/",
      "email": "necanthrope@gmail.com"
    },
    "license": {
      "name": "Apache License Version 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8091",
  "basePath": "/",
...
  "paths": {
...
    "/bookings/addUserToGame": {
      "post": {
        "tags": [
          "bookings-controller"
        ],
        "summary": "Add a registered user to a game",
        "description": "Admin only. \nStatus and Message fields are for output only, and will be ignored on input.",
        "operationId": "addUserToGameUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventBookingData",
            "description": "eventBookingData",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventBookingData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventBookingData"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bookings/bookMeIntoGame": {
      "post": {
        "tags": [
          "bookings-controller"
        ],
        "summary": "Book current user for a game",
        "description": "Book the currently logged in user into a game. Not to be confused with addUserToGame, which is meant to be called by administrators.",
        "operationId": "bookAuthenticatedUserIntoGameUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "gameBooking",
            "description": "gameBooking",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GameBooking"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventBookingReply"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bookings/myAvailableSlots": {
      "get": {
        "tags": [
          "bookings-controller"
        ],
        "summary": "Count of events for a given year",
        "operationId": "getMyCountOfAvailableSlotsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/bookings/removeMeFromGame": {
      "delete": {
        "tags": [
          "bookings-controller"
        ],
        "summary": "Remove current user from a game",
        "description": "Remove the currently logged in user from a game. Not to be confused with removeUserFromGame, which is meant to be called by administrators.",
        "operationId": "removeAuthenticatedUserFromGameUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "gameBooking",
            "description": "gameBooking",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GameBooking"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventBookingReply"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/bookings/removeUserFromGame": {
      "delete": {
        "tags": [
          "bookings-controller"
        ],
        "summary": "Remove a registered user from a game",
        "description": "Admin only. \nStatus and Message fields are for output only, and will be ignored on input.",
        "operationId": "removeUserFromGameUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventBookingData",
            "description": "eventBookingData",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventBookingData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventBookingData"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/bookings/setGmStatusForPlayerInGame": {
      "post": {
        "tags": [
          "bookings-controller"
        ],
        "summary": "Set the GM flag for a user who is booked in a game",
        "description": "Admin only. \nStatus and Message fields are for output only, and will be ignored on input.",
        "operationId": "setGmStatusForPlayerInGameUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "eventBookingData",
            "description": "eventBookingData",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventBookingData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/EventBookingData"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EventBookingReply": {
      "type": "object",
      "properties": {
        "guid": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
    "EventBookingData": {
      "type": "object",
      "properties": {
        "eventId": {
          "type": "integer",
          "format": "int64"
        },
        "isGm": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "GameBooking": {
      "type": "object",
      "properties": {
        "gameId": {
          "type": "integer",
          "format": "int64"
        },
        "uuid": {
          "type": "string"
        }
      }
    },
...
  }
}